Reorganise commands and parsers into student and task packages
