+++ src/main/java/jarvis/logic/commands/student/ClearCommand.java
@@ -1,7 +1,9 @@
-package jarvis.logic.commands;
+package jarvis.logic.commands.student;
 
 import static java.util.Objects.requireNonNull;
 
+import jarvis.logic.commands.Command;
+import jarvis.logic.commands.CommandResult;
 import jarvis.model.StudentBook;
 import jarvis.model.Model;
