+++ src/main/java/jarvis/logic/commands/student/FindCommand.java
@@ -1,8 +1,10 @@
-package jarvis.logic.commands;
+package jarvis.logic.commands.student;
 
 import static java.util.Objects.requireNonNull;
 
 import jarvis.commons.core.Messages;
+import jarvis.logic.commands.Command;
+import jarvis.logic.commands.CommandResult;
 import jarvis.model.Model;
 import jarvis.model.student.NameContainsKeywordsPredicate;
