+++ src/main/java/jarvis/logic/commands/student/ListCommand.java
@@ -1,8 +1,10 @@
-package jarvis.logic.commands;
+package jarvis.logic.commands.student;
 
 import static jarvis.model.Model.PREDICATE_SHOW_ALL_PERSONS;
 import static java.util.Objects.requireNonNull;
 
+import jarvis.logic.commands.Command;
+import jarvis.logic.commands.CommandResult;
 import jarvis.model.Model;
 
 /**
