+++ src/main/java/jarvis/logic/parser/JarvisParser.java
@@ -6,21 +6,25 @@ import static jarvis.commons.core.Messages.MESSAGE_UNKNOWN_COMMAND;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import jarvis.logic.commands.AddCommand;
-import jarvis.logic.commands.ClearCommand;
+import jarvis.logic.commands.student.AddCommand;
+import jarvis.logic.commands.student.ClearCommand;
 import jarvis.logic.commands.Command;
-import jarvis.logic.commands.DeleteCommand;
-import jarvis.logic.commands.EditCommand;
+import jarvis.logic.commands.student.DeleteCommand;
+import jarvis.logic.commands.student.EditCommand;
 import jarvis.logic.commands.ExitCommand;
-import jarvis.logic.commands.FindCommand;
+import jarvis.logic.commands.student.FindCommand;
 import jarvis.logic.commands.HelpCommand;
-import jarvis.logic.commands.ListCommand;
+import jarvis.logic.commands.student.ListCommand;
 import jarvis.logic.parser.exceptions.ParseException;
+import jarvis.logic.parser.student.AddCommandParser;
+import jarvis.logic.parser.student.DeleteCommandParser;
+import jarvis.logic.parser.student.EditCommandParser;
+import jarvis.logic.parser.student.FindCommandParser;
 
 /**
  * Parses user input.
  */
-public class AddressBookParser {
+public class JarvisParser {
 
     /**
      * Used for initial separation of command word and args.
