+++ src/test/java/jarvis/logic/commands/EditCommandTest.java
@@ -14,13 +14,15 @@ import static jarvis.testutil.TypicalPersons.getTypicalAddressBook;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import jarvis.logic.commands.student.ClearCommand;
+import jarvis.logic.commands.student.EditCommand;
 import jarvis.model.StudentBook;
 import jarvis.model.student.Student;
 import org.junit.jupiter.api.Test;
 
 import jarvis.commons.core.Messages;
 import jarvis.commons.core.index.Index;
-import jarvis.logic.commands.EditCommand.EditPersonDescriptor;
+import jarvis.logic.commands.student.EditCommand.EditPersonDescriptor;
 import jarvis.model.Model;
 import jarvis.model.ModelManager;
 import jarvis.model.UserPrefs;
