+++ src/test/java/jarvis/logic/commands/ListCommandTest.java
@@ -5,6 +5,7 @@ import static jarvis.logic.commands.CommandTestUtil.showPersonAtIndex;
 import static jarvis.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 import static jarvis.testutil.TypicalPersons.getTypicalAddressBook;
 
+import jarvis.logic.commands.student.ListCommand;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
