+++ src/test/java/jarvis/logic/parser/DeleteCommandParserTest.java
@@ -5,9 +5,10 @@ import static jarvis.logic.parser.CommandParserTestUtil.assertParseFailure;
 import static jarvis.logic.parser.CommandParserTestUtil.assertParseSuccess;
 import static jarvis.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 
+import jarvis.logic.parser.student.DeleteCommandParser;
 import org.junit.jupiter.api.Test;
 
-import jarvis.logic.commands.DeleteCommand;
+import jarvis.logic.commands.student.DeleteCommand;
 
 /**
  * As we are only doing white-box testing, our test cases do not cover path variations
