+++ src/test/java/jarvis/logic/parser/EditCommandParserTest.java
@@ -31,11 +31,12 @@ import static jarvis.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 import static jarvis.testutil.TypicalIndexes.INDEX_SECOND_PERSON;
 import static jarvis.testutil.TypicalIndexes.INDEX_THIRD_PERSON;
 
+import jarvis.logic.parser.student.EditCommandParser;
 import org.junit.jupiter.api.Test;
 
 import jarvis.commons.core.index.Index;
-import jarvis.logic.commands.EditCommand;
-import jarvis.logic.commands.EditCommand.EditPersonDescriptor;
+import jarvis.logic.commands.student.EditCommand;
+import jarvis.logic.commands.student.EditCommand.EditPersonDescriptor;
 import jarvis.model.student.Address;
 import jarvis.model.student.Email;
 import jarvis.model.student.Name;
